rename "linux" script to "ubuntu"
